move structure into code/__init__.py
